requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.57.0
greenlet==3.2.0
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.54
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.32
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==24.2
pdfkit==1.0.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
PyMuPDF==1.25.5
pymupdf4llm==0.0.21
pyphen==0.17.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
PyPDF2
python-docx
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
zstandard==0.23.0
